% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_timeline.R
\name{draw_panel_function}
\alias{draw_panel_function}
\title{Draw panel}
\usage{
draw_panel_function(data, panel_scales, coord)
}
\arguments{
\item{data}{the data.frame from the geom_timeline function call}

\item{panel_scales}{the panel_scales}

\item{coord}{the modified data.frame labeled by the parameters of aes()}
}
\value{
an object to be used in geom_timeline
}
\description{
This further modifies the provided data for mapping within geom_timeline
}
\examples{
\dontrun{
ggplot(clean_earthquake_data, aes(
x = date,
y = COUNTRY,
color = TOTAL_DEATHS,
size = EQ_PRIMARY
),
n_max = 10) +
geom_timeline()
}

}
