% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthquake_read.R
\name{earthquake_read}
\alias{earthquake_read}
\title{Read earthquake data}
\usage{
earthquake_read()
}
\value{
A dplyr tbl_df of the raw earthquake data
}
\description{
This function reads in the earthquake file and turns into a dplyr tbl_df
}
\examples{
\dontrun{
earthquake_read()
}

}
