% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_clean_data.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{Clean earthquake data}
\usage{
eq_clean_data(raw_data)
}
\arguments{
\item{raw_data}{The raw earthquake data.frame}
}
\value{
A cleaned earthquake data.frame
}
\description{
This function cleans raw earthquake data. It creates a date column, cleans
the location column, and converts the latitute and longtitude columns to numeric
}
\note{
the function will throw an error if raw_data is not a data.frame
}
\examples{
\dontrun{
eq_clean_data(raw_earthquake_data)
}

}
