% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_create_label.R
\name{eq_create_label}
\alias{eq_create_label}
\title{Create a Leaflet map label}
\usage{
eq_create_label(x)
}
\arguments{
\item{x}{A clean earthquake data.frame}
}
\value{
The clean earthquake data.frame with a popup_text variable
}
\description{
This function builds a label for the leaflet map popup based off of the
location, magnitude, and death toll of the earthquake represented in the
data
}
\note{
the function will throw an error if x is not a data.frame
}
\examples{
\dontrun{
clean_earthquake_data$popup_text <- eq_create_label(clean_earthquake_data)
}

}
