% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_location_clean.R
\name{eq_location_clean}
\alias{eq_location_clean}
\title{Clean earthquake data location column}
\usage{
eq_location_clean(raw_data1)
}
\arguments{
\item{raw_data1}{The raw earthquake data.frame}
}
\value{
An earthquake data.frame with a cleaned LOCATION_NAME column
}
\description{
This function cleans the raw earthquake data location column, removing the
country part of the string
}
\note{
the function will throw an error if raw_data is not a data.frame
}
\examples{
\dontrun{
eq_location_clean(raw_earthquake_data)
}

}
