% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_map.R
\name{eq_map}
\alias{eq_map}
\title{Create a leaflet map of earthquake data}
\usage{
eq_map(x, annot_col)
}
\arguments{
\item{x}{a clean earthquake data.frame}

\item{annot_col}{a character vector from the earthquake data.frame used to label the popups on the map}
}
\value{
A leaflet map
}
\description{
This function reads in a given file and turns into a table.
}
\note{
the function will throw an error if x is not a data.frame

the function will throw an error if annot_col is not a non-null column of x
}
\examples{
\dontrun{
eq_map(clean_earthquake_data, "date")
eq_map(cleaned_earthquake_data, "TOTAL_DEATHS")
}

}
