% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_timeline.R
\name{geom_timeline}
\alias{geom_timeline}
\title{geom_timeline}
\usage{
geom_timeline(mapping = NULL, data = NULL, stat = "timeline",
  position = "identity", show.legend = NA, na.rm = FALSE,
  inherit.aes = TRUE, ...)
}
\arguments{
\item{mapping}{NULL}

\item{data}{NULL}

\item{stat}{timeline}

\item{position}{identity}

\item{show.legend}{NA}

\item{na.rm}{FALSE}

\item{inherit.aes}{TRUE}

\item{...}{any other arguments passed to geom_timeline}
}
\value{
a ggplot object
}
\description{
This actually does the plotting
}
\examples{
\dontrun{
ggplot(clean_earthquake_data, aes(
x = date,
y = COUNTRY,
color = TOTAL_DEATHS,
size = EQ_PRIMARY
),
n_max = 10) +
geom_timeline()
}

}
