% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_timeline.R
\docType{data}
\name{GeomTimeline}
\alias{GeomTimeline}
\title{GeomTimeline}
\format{An object of class \code{GeomTimeline} (inherits from \code{GeomPoint}, \code{Geom}, \code{ggproto}) of length 5.}
\usage{
GeomTimeline
}
\value{
a geom
}
\description{
This function maps user-provided earthquake data
}
\examples{
\dontrun{
ggplot(clean_earthquake_data, aes(
x = date,
y = COUNTRY,
color = TOTAL_DEATHS,
size = EQ_PRIMARY
),
n_max = 10) +
geom_timeline()
}

}
\keyword{datasets}
