% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_timeline_label.R
\docType{data}
\name{GeomTimelineLabel}
\alias{GeomTimelineLabel}
\title{GeomTimelineLabel}
\format{An object of class \code{GeomTimelineLabel} (inherits from \code{GeomPoint}, \code{Geom}, \code{ggproto}) of length 5.}
\usage{
GeomTimelineLabel
}
\value{
a geom
}
\description{
This function maps user-provided earthquake data
}
\examples{
\dontrun{
ggplot(clean_earthquake_data, aes(
x = date,
y = COUNTRY,
color = TOTAL_DEATHS,
size = EQ_PRIMARY,
label = LOCATION_NAME
),
n_max = 10) +
geom_timeline_label()
}

}
\keyword{datasets}
