% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_timeline_label.R
\name{label_draw_panel_function}
\alias{label_draw_panel_function}
\title{Label Draw panel}
\usage{
label_draw_panel_function(data, panel_scales, coord)
}
\arguments{
\item{data}{the data.frame from the geom_timeline_label function call}

\item{panel_scales}{the panel_scales}

\item{coord}{the modified data.frame labeled by the parameters of aes()}
}
\value{
an object to be used in geom_timeline_label
}
\description{
This further modifies the provided data for mapping within geom_timeline_label
}
\examples{
\dontrun{
ggplot(clean_earthquake_data, aes(
x = date,
y = COUNTRY,
color = TOTAL_DEATHS,
size = EQ_PRIMARY,
label = LOCATION_NAME
),
n_max = 10) +
geom_timeline_label()
}

}
