% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_timeline_label.R
\name{stat_timeline_label}
\alias{stat_timeline_label}
\title{stat_timeline_label}
\usage{
stat_timeline_label(mapping = NULL, data = NULL, geom = "timeline_label",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, n_max = Inf, ...)
}
\arguments{
\item{mapping}{NULL}

\item{data}{NULL}

\item{geom}{polygon}

\item{position}{identity}

\item{na.rm}{FALSE}

\item{show.legend}{NA}

\item{inherit.aes}{TRUE}

\item{n_max}{the maximum number of earthquakes to be plotted}

\item{...}{any other arguments passed to stat_timeline_label}
}
\value{
a stat
}
\description{
This is the actual stat that prepares data for mapping by geom_timeline
}
\examples{
\dontrun{
ggplot(clean_earthquake_data, aes(
x = date,
y = COUNTRY,
color = TOTAL_DEATHS,
size = EQ_PRIMARY,
label = LOCATION_NAME
),
n_max = 10) +
geom_timeline_label()
}

}
