% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_timeline.R
\docType{data}
\name{StatTimeline}
\alias{StatTimeline}
\title{Timeline Stat}
\format{An object of class \code{StatTimeline} (inherits from \code{Stat}, \code{ggproto}) of length 3.}
\usage{
StatTimeline
}
\value{
a stat
}
\description{
This stat prepares data for mapping by geom_timeline by subsetting the data
to the user-provided minimum and maximum dates
}
\examples{
\dontrun{
ggplot(clean_earthquake_data, aes(
x = date,
y = COUNTRY,
color = TOTAL_DEATHS,
size = EQ_PRIMARY
),
n_max = 10) +
geom_timeline()
}

}
\keyword{datasets}
