% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_timeline_label.R
\docType{data}
\name{StatTimelineLabel}
\alias{StatTimelineLabel}
\title{Timeline Label Stat}
\format{An object of class \code{StatTimeline} (inherits from \code{Stat}, \code{ggproto}) of length 3.}
\usage{
StatTimelineLabel
}
\value{
a stat
}
\description{
This stat prepares data for mapping by geom_timeline by subsetting the data
to the user-provided minimum and maximum dates as well as to the number of
earthquakes specified by the user in the n_max variable
}
\examples{
\dontrun{
ggplot(clean_earthquake_data, aes(
x = date,
y = COUNTRY,
color = TOTAL_DEATHS,
size = EQ_PRIMARY,
label = LOCATION_NAME
),
n_max = 10) +
geom_timeline_label()
}

}
\keyword{datasets}
